feat: details page completed.

+++ core/model/src/main/java/com/github/smmousavi/model/Planet.kt
@@ -3,5 +3,5 @@ package com.github.smmousavi.model
 data class Planet(
     val id: String,
     val name: String,
-    val population: Int,
+    val population: Double,
 )
